edits to routes and models to work better with CLI

+++ app/routes.py
@@ -32,6 +32,14 @@ def video_not_found(func):
     inner.__name__ = func.__name__
     return inner
 
+def rental_not_found(func):
+    def inner(rental_id):
+        if Rental.query.get(rental_id) is None:
+            return jsonify({"details": "rental not found"}), 404
+        return func(rental_id)
+    inner.__name__ = func.__name__
+    return inner
+
 #---------------------# CUSTOMER ENDPOINTS #---------------------#
 
 @customers_bp.route("", methods=["GET"], strict_slashes=False)
@@ -55,7 +63,6 @@ def create_customer():
                     postal_code = str(request_body["postal_code"]),
                     phone = str(request_body["phone"]),
                     registered_at = datetime.now(),
-                    #modify this line or Customers model to actually work (but in theory a new customer added should have a video_out_count of 0)
                     videos_checked_out_count = 0)
     db.session.add(new_customer)
     db.session.commit()
@@ -78,24 +85,30 @@ def update_customer(customer_id):
 @customer_not_found
 def delete_customer(customer_id):
     customer = Customer.query.get(customer_id)
+    if customer.videos_checked_out_count:
+        return jsonify({"details": "Invalid Request: This customer has an active rental and cannot be deleted."})
+    for rental in customer.rentals:
+        db.session.delete(rental)
     db.session.delete(customer)
     db.session.commit()
-    return jsonify({"details":"customer successfully deleted",
+    return jsonify({"details":f"Customer '{customer.name}' successfully deleted.",
                     "id": customer.id,
                     "name": customer.name}), 200
 
 @customers_bp.route("/<customer_id>/rentals", methods=["GET"], strict_slashes=False)
 @customer_not_found
-def get_customer_rentals(customer_id):
+def get_current_customer_rentals(customer_id):
     customer = Customer.query.get(customer_id)
     rentals = []
-    for rental in customer.videos:
-        video = Video.query.get(rental.video_id)
-        dict = {
-            "title": video.title,
-            "release_date": video.release_date,
-            "due_date": rental.due_date}
-        rentals.append(dict)
+    for rental in customer.rentals:
+        if rental.check_in_date == None:
+            video = Video.query.get(rental.video_id)
+            dict = {
+                "title": video.title,
+                "release_date": video.release_date,
+                "due_date": rental.due_date,
+                "check_in_date": rental.check_in_date}
+            rentals.append(dict)
     return jsonify(rentals), 200
 
 
@@ -135,7 +148,9 @@ def update_video(video_id):
             return invalid_input()
     video.title = response_body["title"]
     video.release_date = response_body["release_date"]
+    video.available_inventory = response_body["total_inventory"] - (video.total_inventory - video.available_inventory)
     video.total_inventory = response_body["total_inventory"]
+
     db.session.commit()
     return jsonify(video.to_json()), 200
 
@@ -143,25 +158,31 @@ def update_video(video_id):
 @video_not_found
 def delete_video(video_id):
     video = Video.query.get(video_id)
+    if video.available_inventory != video.total_inventory:
+        return jsonify({"details": "Invalid Request: This video has an active rental and cannot be deleted."})
+    for rental in video.rentals:
+        db.session.delete(rental)
     db.session.delete(video)
     db.session.commit()
-    return jsonify({"details":"video successfully deleted",
+    return jsonify({"details":f"Video '{video.title}' successfully deleted.",
                     "id": video.id,
                     "title": video.title}), 200
 
 @videos_bp.route("/<video_id>/rentals", methods=["GET"], strict_slashes=False)
 @video_not_found
-def get_video_rentals(video_id):
+def get_current_video_rentals(video_id):
     video = Video.query.get(video_id)
     rentals = []
-    for rental in video.customers:
-        customer = Customer.query.get(rental.customer_id)
-        dict = {
-            "name": customer.name,
-            "phone": customer.phone,
-            "postal_code": customer.postal_code,
-            "due_date": rental.due_date}
-        rentals.append(dict)
+    for rental in video.rentals:
+        if rental.check_in_date == None:
+            customer = Customer.query.get(rental.customer_id)
+            dict = {
+                "name": customer.name,
+                "phone": customer.phone,
+                "postal_code": customer.postal_code,
+                "due_date": rental.due_date,
+                "check_in_date": rental.check_in_date}
+            rentals.append(dict)
     return jsonify(rentals), 200
 
 
@@ -174,14 +195,12 @@ def check_out_video():
     video_id = request_body["video_id"]
     if not isinstance(video_id, int) or not isinstance(customer_id, int):
         return invalid_input()
-
     customer = Customer.query.get(customer_id)
     video = Video.query.get(video_id)
     if customer is None or video is None:
         return jsonify({"details": "customer or video not Found"}), 404
     if video.available_inventory == 0:
         return jsonify({"details": "Video out of stock"}), 400
-
     customer.videos_checked_out_count += 1
     video.available_inventory -= 1
     new_rental = Rental(customer_id = customer_id,
@@ -190,7 +209,7 @@ def check_out_video():
                     checkout_date = datetime.now())
     db.session.add(new_rental)
     db.session.commit()
-    return jsonify(new_rental.check_out_to_json()), 200
+    return jsonify(new_rental.check_rental_to_json()), 200
 
 @rentals_bp.route("/check-in", methods=["POST"], strict_slashes=False)
 def check_in_video():
@@ -198,13 +217,15 @@ def check_in_video():
     request_body = request.get_json()
     customer_id = request_body["customer_id"]
     video_id = request_body["video_id"]
+    customer = Customer.query.get(customer_id)
+    video = Video.query.get(video_id)
     for rental in rentals:
         if rental.customer_id == customer_id and rental.video_id == video_id and rental.check_in_date == None:
-            rental.customer.videos_checked_out_count -= 1
-            rental.video.available_inventory += 1
+            customer.videos_checked_out_count -= 1
+            video.available_inventory += 1
             rental.check_in_date = datetime.now()
             db.session.commit()
-            return jsonify(rental.check_in_to_json()), 200
+            return jsonify(rental.check_rental_to_json()), 200
     return invalid_input()
 
 
@@ -220,7 +241,7 @@ def overdue_rentals_index():
 @video_not_found
 def video_history(video_id):
     video = Video.query.get(video_id)
-    rentals = video.customers
+    rentals = video.rentals
     video_history = [rental.video_history_to_json() for rental in rentals if rental.check_in_date]
     return jsonify(video_history), 200
 
@@ -228,6 +249,6 @@ def video_history(video_id):
 @customer_not_found
 def customer_history(customer_id):
     customer = Customer.query.get(customer_id)
-    rentals = customer.videos
+    rentals = customer.rentals
     customer_history = [rental.customer_history_to_json() for rental in rentals if rental.check_in_date]
     return jsonify(customer_history), 200
